refactor: move Notification type to a shared types file and update imports in Notifications component
