+++ apps/erp/app/types/index.ts
@@ -29,6 +29,23 @@ export type ListItem = {
   name: string;
 };
 
+// Topbar shape — matches the surface the old Novu IMessage exposed so
+// consumers don't change. `payload` mirrors what notify.ts writes into the
+// notification row's payload jsonb column.
+export type Notification = {
+  _id: string;
+  read: boolean;
+  seen: boolean;
+  createdAt: string;
+  payload: {
+    recordId?: string;
+    description?: string;
+    event?: string;
+    from?: string;
+    documentType?: string;
+  };
+};
+
 export type ModelUpload = {
   modelId: string | null;
   modelName: string | null;
