+++ packages/ee/src/plan.ts
@@ -6,12 +6,11 @@ import type { IntegrationID } from "./index";
  * (`usePlanGate`) and server (`plan.server.ts`) read from here.
  */
 export const FEATURE_PLANS = {
-  API_KEYS: [Plan.Starter, Plan.Business],
-  AUDIT_LOG: [Plan.Business],
-  EMAIL_NOTIFICATIONS: [Plan.Business],
-  INTEGRATIONS: [Plan.Starter, Plan.Business],
+  API_KEYS: [Plan.Business, Plan.Partner],
+  WEBHOOKS: [Plan.Business, Plan.Partner],
+  INTEGRATIONS: [Plan.Business, Plan.Partner],
   ITEM_RULES: [Plan.Business],
-  WEBHOOKS: [Plan.Starter, Plan.Business]
+  AUDIT_LOG: [Plan.Business]
 } as const satisfies Record<string, Plan[]>;
 
 export type Feature = keyof typeof FEATURE_PLANS;
